% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{camel_to_title}
\alias{camel_to_title}
\title{camel_to_title}
\usage{
camel_to_title(camel_case)
}
\arguments{
\item{camel_case}{input character string}
}
\value{
string
}
\description{
Convert character string from camel case to title/proper case
}
