% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codmw-package.R
\docType{package}
\name{codmw-package}
\alias{codmw}
\alias{codmw-package}
\title{codmw: Call of Duty Modern Warfare Customized Shiny Application}
\description{
CODMW is an R package and Shiny app that extracts,
    tracks, stores, and facilitates Call of Duty statistics in real time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig2011/codmw}
  \item Report bugs at \url{https://github.com/jimbrig2011/codmw/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

Other contributors:
\itemize{
  \item Mac Kitchen \email{w.mac.kitchin@gmail.com} [contributor]
  \item The Documentalarialists Foundation, Inc. [funder]
}

}
\keyword{internal}
