% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flucol}
\alias{flucol}
\title{Fluid Column - Shiny fluidRow + Column}
\usage{
flucol(..., width = 12, offset = 0)
}
\arguments{
\item{...}{elements to include within the flucol}

\item{width}{width}

\item{offset}{offset}
}
\value{
A column wrapped in fluidRow
}
\description{
Fluid Column - Shiny fluidRow + Column
}
\examples{
propertyAllocation::flucol(12, 0, shiny::h5("HEY"))
}
