% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(gamertag)
}
\arguments{
\item{gamertag}{xbox live gamertag}
}
\value{
tibble
}
\description{
retrieve data from \url{https://my.callofduty.com} API by xbox gamertag
}
\examples{
get_data(munchinxbox69)

}
