% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url}
\alias{get_url}
\title{Get URL}
\usage{
get_url(gamertag)
}
\arguments{
\item{gamertag}{gamertag (string)}
}
\value{
character vector of URLS
}
\description{
Retrives URL specific to a certain gamertag
}
