% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icon_text}
\alias{icon_text}
\title{Icon Text}
\usage{
icon_text(icon, text)
}
\arguments{
\item{icon}{fontawesome icon}

\item{text}{text}
}
\value{
HTML div
}
\description{
Creates an HTML div containing the icon and text.
}
\examples{
icon_text("table", "Table")

}
