% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_logo}
\alias{insert_logo}
\title{Insert Logo}
\usage{
insert_logo(
  file,
  style = "background-color: #FFF; width: 100\%; height: 100\%;",
  width = NULL,
  ref = "#"
)

insert_logo(
  file,
  style = "background-color: #FFF; width: 100\%; height: 100\%;",
  width = NULL,
  ref = "#"
)
}
\arguments{
\item{file}{file}

\item{style}{style}

\item{width}{width}

\item{ref}{ref}
}
\value{


tag
}
\description{
Insert Logo

Insert Logo
}
