% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codmw-package.R
\docType{package}
\name{codmw-package}
\alias{codmw}
\alias{codmw-package}
\title{codmw: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

}
\keyword{internal}
